1.4.0
Added a new item in the settings, enabled by default when first installing the script: Add button to notifications in the /account/alerts list. When enabled, adds button to notifications.
Added new item in the settings, enabled by default on the first installation of the script: Add button to dialogs in personal messages. When enabled, adds button to dialogs:
Fixed when checking your profile in https://zelenka.guru/online/?type=registered would give you an error. Now your profile is not checked

Added new items to the statistics table:

"Don't fit the specified parameters" (is added only if automatic checking only by parameters is enabled in settings, the item is not added on the page with users who have recently registered) - displays the number of users who don't fit the specified parameters (likes and messages more than specified in the parameters)

"Check left" - displays the number of users left on the page for checking.

Translated with www.DeepL.com/Translator (free version)